requirements.txt
streamlit
numpy
pandas
matplotlib
python-dotenv
requests
sympy
scipy
vedo
Pillow
numpy-stl
plotly
pythonocc-core